+++ BookRecommendationSystem/BookRecommendationSystem/settings.py
@@ -39,7 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'myapp', 
-    'bst_search'
+    'search'
 ]
 
 MIDDLEWARE = [
