+++ BookRecommendationSystem/search/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class BstSearchConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'bst_search'
+    name = 'search'
