+++ BookRecommendationSystem/search/bst.py
@@ -5,8 +5,9 @@ class Node:
         self.right = None
 
 class BookBST:
-    def __init__(self):
+    def __init__(self, key):
         self.root = None
+        self.key = key
     
     def insert(self, book):
         """Insert a book into the BST"""
@@ -16,7 +17,9 @@ class BookBST:
         
         current = self.root
         while True:
-            if book.title.lower() < current.book.title.lower():
+            book_key = getattr(book, self.key).lower()
+            current_key = getattr(current.book, self.key).lower()
+            if book_key < current_key:
                 if current.left is None:
                     current.left = Node(book)
                     break
@@ -26,6 +29,16 @@ class BookBST:
                     current.right = Node(book)
                     break
                 current = current.right
+            #if book.title.lower() < current.book.title.lower():
+            #    if current.left is None:
+            #        current.left = Node(book)
+            #        break
+            #    current = current.left
+            #else:
+            #    if current.right is None:
+            #        current.right = Node(book)
+            #        break
+            #    current = current.right
     
     def search(self, query, limit=25):
         """Search for books with titles containing the query"""
@@ -40,7 +53,8 @@ class BookBST:
             inorder_search(node.left)
             
             # Process current node
-            if len(results) < limit and query in node.book.title.lower():
+            book_key = getattr(node.book, self.key).lower()
+            if len(results) < limit and query in book_key:
                 results.append(node.book)
             
             # Traverse right subtree
